+++ dbWave2/Envelope.cpp
@@ -14,13 +14,13 @@ IMPLEMENT_SERIAL(CHighLight, CObject, 0 /* schema number*/)
 CHighLight::CHighLight()
 = default;
 
-void CHighLight::Serialize(CArchive & ar)
+void CHighLight::Serialize(CArchive& ar)
 {
 	if (ar.IsStoring())
 	{
 		const WORD version = 1;
-		ar << version;				// 1
-		const auto nitems = static_cast<int>(3);
+		ar << version; // 1
+		const auto nitems = 3;
 		ar << nitems;
 		ar << channel;
 		ar << color;
@@ -40,9 +40,10 @@ void CHighLight::Serialize(CArchive & ar)
 	l_last.Serialize(ar);
 }
 
-CHighLight& CHighLight::operator=(const CHighLight & arg)
+CHighLight& CHighLight::operator=(const CHighLight& arg)
 {
-	if (this != &arg) {
+	if (this != &arg)
+	{
 		channel = arg.channel;
 		color = arg.color;
 		pensize = arg.pensize;
@@ -56,26 +57,26 @@ IMPLEMENT_SERIAL(CEnvelope, CObject, 0 /* schema number*/)
 
 CEnvelope::CEnvelope()
 {
-	m_sourceMode = 0;			// operation on raw data (nop, deriv, ...)
-	m_sourceChan = 0;			// source channel
-	m_dataperpixel = 2;		// Envelope w. max/min
+	m_sourceMode = 0; // operation on raw data (nop, deriv, ...)
+	m_sourceChan = 0; // source channel
+	m_dataperpixel = 2; // Envelope w. max/min
 	m_npixels = 0;
 	m_span = 0;
 }
 
 CEnvelope::CEnvelope(WORD npixels)
 {
-	m_dataperpixel = 2;		// Envelope w. max/min
+	m_dataperpixel = 2; // Envelope w. max/min
 	m_npixels = npixels;
 	m_Envelope.SetSize(m_npixels, m_dataperpixel);
-	m_sourceMode = 0;			// operation on raw data (nop, deriv, ...)
-	m_sourceChan = 0;			// source channel
+	m_sourceMode = 0; // operation on raw data (nop, deriv, ...)
+	m_sourceChan = 0; // source channel
 	m_span = 0;
 }
 
 CEnvelope::CEnvelope(WORD npixels, int dataperpixel, int ns, int mode, int span)
 {
-	m_dataperpixel = dataperpixel;		// Envelope w. max/min
+	m_dataperpixel = dataperpixel; // Envelope w. max/min
 	m_npixels = npixels;
 	m_Envelope.SetSize(m_npixels * m_dataperpixel);
 	m_sourceMode = mode;
@@ -146,37 +147,37 @@ void CEnvelope::FillEnvelopeWithAbcissaEx(int pixfirst, int pixlast, int ndatapo
 	}
 }
 
-void CEnvelope::ExportToAbcissa(CArray<CPoint, CPoint> &dest)
+void CEnvelope::ExportToAbcissa(CArray<CPoint, CPoint>& dest)
 {
 	for (auto i = 0; i < m_Envelope.GetSize(); i++)
 		dest.GetAt(i).x = m_Envelope[i];
 }
 
-void CEnvelope::ExportToOrdinates(CArray<CPoint, CPoint> &dest)
+void CEnvelope::ExportToOrdinates(CArray<CPoint, CPoint>& dest)
 {
 	for (auto i = 0; i < m_Envelope.GetSize(); i++)
 		dest.GetAt(i).y = m_Envelope[i];
 }
 
-void CEnvelope::Serialize(CArchive & ar)
+void CEnvelope::Serialize(CArchive& ar)
 {
-	m_Envelope.Serialize(ar);	// Envelope array
+	m_Envelope.Serialize(ar); // Envelope array
 
 	if (ar.IsStoring())
 	{
-		ar << m_sourceMode;		// operation on raw data (nop, deriv, ...)
-		ar << m_sourceChan;		// source channel
-		ar << m_span;			// additionnal data pts necessary to compute transform
-		ar << m_dataperpixel;	// 2=simple envelope, 1=raw data
-		ar << m_npixels;		// nb of valid elements in the array
+		ar << m_sourceMode; // operation on raw data (nop, deriv, ...)
+		ar << m_sourceChan; // source channel
+		ar << m_span; // additionnal data pts necessary to compute transform
+		ar << m_dataperpixel; // 2=simple envelope, 1=raw data
+		ar << m_npixels; // nb of valid elements in the array
 	}
 	else
 	{
-		ar >> m_sourceMode;		// operation on raw data (nop, deriv, ...)
-		ar >> m_sourceChan;		// source channel
-		ar >> m_span;			// additionnal data pts necessary to compute transform
-		ar >> m_dataperpixel;	// 2=simple envelope, 1=raw data
-		ar >> m_npixels;		// nb of valid elements in the array
+		ar >> m_sourceMode; // operation on raw data (nop, deriv, ...)
+		ar >> m_sourceChan; // source channel
+		ar >> m_span; // additionnal data pts necessary to compute transform
+		ar >> m_dataperpixel; // 2=simple envelope, 1=raw data
+		ar >> m_npixels; // nb of valid elements in the array
 	}
 }
 
@@ -184,25 +185,25 @@ void CEnvelope::Serialize(CArchive & ar)
 // copy average of 2 consecutive pts (ie (max + min)/2)
 // used by lineview to indicate where curves are dragged by the mouse
 
-void CEnvelope::GetMeanToAbcissa(CArray<CPoint, CPoint> &dest)
+void CEnvelope::GetMeanToAbcissa(CArray<CPoint, CPoint>& dest)
 {
-	auto lp_source = &m_Envelope[0];	// source data: a Envelope
+	auto lp_source = &m_Envelope[0]; // source data: a Envelope
 	auto jdest = 0;
 	for (auto i = m_Envelope.GetSize() / 2; i > 0; i--)
 	{
-		dest[jdest].x = (*lp_source / 2 + *(lp_source + 1) / 2);	// copy half value
+		dest[jdest].x = (*lp_source / 2 + *(lp_source + 1) / 2); // copy half value
 		jdest++;
 		lp_source += 2;
 	}
 }
 
-void CEnvelope::GetMeanToOrdinates(CArray<CPoint, CPoint> &dest)
+void CEnvelope::GetMeanToOrdinates(CArray<CPoint, CPoint>& dest)
 {
-	auto lp_source = &m_Envelope[0];	// source data: a Envelope
+	auto lp_source = &m_Envelope[0]; // source data: a Envelope
 	auto jdest = 0;
 	for (auto i = m_Envelope.GetSize() / 2; i > 0; i--)
 	{
-		dest[jdest].y = (*lp_source / 2 + *(lp_source + 1) / 2);	// copy half value
+		dest[jdest].y = (*lp_source / 2 + *(lp_source + 1) / 2); // copy half value
 		jdest++;
 		lp_source += 2;
 	}
@@ -224,34 +225,35 @@ void CEnvelope::SetEnvelopeSize(int npixels, int ndataperpixel)
 void CEnvelope::FillEnvelopeWithMxMi(int ifirst, short* lp_data, int nchans, int nelmts, BOOL b_new)
 {
 	auto lp_envelope = &m_Envelope[ifirst * m_dataperpixel];
-	if (m_dataperpixel == 1)	// only one data point per pixel
+	if (m_dataperpixel == 1) // only one data point per pixel
 	{
-		*lp_envelope = *lp_data;	// simply copy data
+		*lp_envelope = *lp_data; // simply copy data
 		return;
 	}
 
-	int i_max = *lp_data;			// init max
-	auto i_min = i_max;				// init min
+	int i_max = *lp_data; // init max
+	auto i_min = i_max; // init min
 	if (!b_new)
 	{
 		i_min = *lp_envelope;
 		i_max = *(lp_envelope + 1);
 	}
 
-	while (nelmts > 0)			// scan nb elements
-	{							// designed by scale
-		const auto idata = *lp_data;		// load value
+	while (nelmts > 0) // scan nb elements
+	{
+		// designed by scale
+		const auto idata = *lp_data; // load value
 		if (idata < i_min)
-			i_min = idata;		// change min
+			i_min = idata; // change min
 		else if (idata > i_max)
-			i_max = idata;		// change max
-		lp_data += nchans;		// update data pointer
+			i_max = idata; // change max
+		lp_data += nchans; // update data pointer
 		nelmts--;
 	}
 
-	*lp_envelope = i_max;		// store max
-	lp_envelope++;				// update Envelope pointer
-	*lp_envelope = i_min;		// store min
+	*lp_envelope = i_max; // store max
+	lp_envelope++; // update Envelope pointer
+	*lp_envelope = i_min; // store min
 }
 
 short* CEnvelope::getMaxMin(int nelements, short* lpData, int nchans, short& i_min, short& i_max, long& y1)
@@ -278,17 +280,17 @@ void CEnvelope::FillEnvelopeWithSmoothMxMi(int ifirst, short* lpData, int nchans
 {
 	// simple case: save just the raw data
 	short* lp_envelope = &m_Envelope[ifirst * m_dataperpixel];
-	if (m_dataperpixel == 1)	// only one data point per pixel
+	if (m_dataperpixel == 1) // only one data point per pixel
 	{
-		*lp_envelope = *lpData;	// simply copy data
+		*lp_envelope = *lpData; // simply copy data
 		return;
 	}
 
 	// more than 1 point per interval
 	// cut interval in 2 and store either [max, min] or [min, max] depending on the direction of the curve
 	// find direction by comparing the average point between the 2 sub-intervals
-	short i_max = *lpData;		// init max
-	short i_min = i_max;		// init min
+	short i_max = *lpData; // init max
+	short i_min = i_max; // init min
 	if (!bNew)
 	{
 		i_min = *lp_envelope;
@@ -316,8 +318,8 @@ void CEnvelope::FillEnvelopeWithSmoothMxMi(int ifirst, short* lpData, int nchans
 	}
 	else //if (ioption == 2)
 	{
-		i_min = (short)y1;
-		i_max = (short)y2;
+		i_min = static_cast<short>(y1);
+		i_max = static_cast<short>(y2);
 	}
 
 	*lp_envelope = i_min;
@@ -342,8 +344,8 @@ void CEnvelope::GetEnvelopeMaxMin(int* max, int* min)
 		if (val < minval)
 			minval = val;
 	}
-	*min = (int)minval;
-	*max = (int)maxval;
+	*min = static_cast<int>(minval);
+	*max = static_cast<int>(maxval);
 }
 
 void CEnvelope::GetEnvelopeMaxMinBetweenPoints(int ifirstpixel, int ilastpixel, int* max, int* min)
@@ -358,6 +360,6 @@ void CEnvelope::GetEnvelopeMaxMinBetweenPoints(int ifirstpixel, int ilastpixel, 
 		if (val < minval)
 			minval = val;
 	}
-	*min = (int)minval;
-	*max = (int)maxval;
+	*min = static_cast<int>(minval);
+	*max = static_cast<int>(maxval);
 }
